requirements.txt
pyupbit>=0.2.33
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT>=2.8.0
schedule>=1.2.0
pytest>=7.4.0
